@@ -18,23 +18,23 @@
         <module>jc-file</module>
     </modules>
     <properties>
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>17</java.version>
-        <spring.cloud.alibaba.version>2021.0.5.0</spring.cloud.alibaba.version>
-        <spring.cloud.version>2021.0.5</spring.cloud.version>
-        <spring.boot.version>2.6.13</spring.boot.version>
+        <java.version>21</java.version>
+        <spring.cloud.alibaba.version>2023.0.0.0-RC1</spring.cloud.alibaba.version>
+        <spring.cloud.version>2023.0.1</spring.cloud.version>
+        <spring.boot.version>3.2.4</spring.boot.version>
         <lombok>1.18.28</lombok>
         <jackson>2.16.1</jackson>
         <redisson>3.17.1</redisson>
         <mysql-connector.version>8.0.31</mysql-connector.version>
-        <mybatis-plus.version>3.5.4</mybatis-plus.version>
+        <mybatis-plus.version>3.5.6</mybatis-plus.version>
         <hutool>5.8.26</hutool>
         <juint>4.13.2</juint>
         <ipfs>1.3.3</ipfs>
         <log>1.4.7</log>
-        <dubbo>3.1.2</dubbo>
+        <dubbo>3.3.0-beta.2</dubbo>
     </properties>
     <dependencyManagement>
         <dependencies>
