@@ -2,7 +2,7 @@ package www.raven.jc.util;
 
 import www.raven.jc.constant.JwtConstant;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 
 /**
