@@ -2,7 +2,7 @@ package www.raven.jc.ws;
 
 import www.raven.jc.entity.dto.MessageDTO;
 
-import javax.websocket.Session;
+import jakarta.websocket.Session;
 
 
 /**
